refactor core
- refactoring core to provide more code seperation
- move Kprobe/Uprobe/Tracepoint attach into constructor for those structs
